# Parallel UDF optimizations