@dreadatour 2024-08-07
> Thank you, Vlad! Great improvements.

Thank you ❤️

> 1. **Is there a way to avoid the conversion?** We use a very limited subset of types. Why not just pass unflattened objects as tuple?
> 2. **Is there a way to avoid marshaling whatsoever?** We can write directly to DB.
> 
> Think of these as ideas for the next steps.

Yes, I have quite a lot of ideas to improve things more, these suggestions looks great! Let me put it all together in the optimizations epic 👍